requirements.txt
behave>=1.2.6
python-dotenv==1.2.1
pytest-playwright==0.7.0
playwright==1.54.0
pyhamcrest==2.1.0
junit2html==31.0.5
behave-html-formatter==0.9.10
behavex==4.4.2
jsonschema==4.23.0
dominate==2.9.1
requests==2.32.4
typing_extensions==4.14.1
colorama==0.4.6
Pygments==2.19.2
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
pytest==8.4.2
boto3==1.41.3
psycopg2-binary==2.9.11
pandas==2.3.3
pymysql==1.1.2
openpyxl==3.1.5